Add function to test non-factory creation
