+++ tests/Unit/Time/TimeSteppers/Test_AdamsBashforthN.cpp
@@ -13,7 +13,7 @@
 #include "Parallel/RegisterDerivedClassesWithCharm.hpp"
 #include "Time/TimeSteppers/AdamsBashforthN.hpp"
 #include "Utilities/Gsl.hpp"
-#include "tests/Unit/TestFactoryCreation.hpp"
+#include "tests/Unit/TestCreation.hpp"
 #include "tests/Unit/TestHelpers.hpp"
 #include "tests/Unit/Time/TimeSteppers/TimeStepperTestUtils.hpp"
